+++ test/is-stale-test-def.ts
@@ -1,10 +1,10 @@
-import { StaleMap } from "../src/stale";
+import { FieldsMap } from "graphql-norm";
 
 export interface OneTest {
   readonly name: string;
   readonly only?: boolean;
   readonly skip?: boolean;
   readonly fields: { readonly [key: string]: ReadonlyArray<string> };
-  readonly staleMap: StaleMap;
+  readonly staleMap: FieldsMap;
   readonly isStale: boolean;
 }