+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased](https://github.com/dividab/graphql-norm-stale/compare/v0.3.0...master)
 
+### Changed
+
+- Use the `FieldsMap` type from `graphql-norm` instead of `StaleMap` and `StaleFields`.
+- Rename `updateStale()` to `clearStale()`.
+
 ## [0.2.0](https://github.com/dividab/graphql-norm-stale/compare/v0.2.0...v0.3.0) - 2019-10-02
 
 ### Added